houdini: whilte list Andebench in x86 native lib mismatch issue
BZ: 186161

1. In the beginning of the long discussion about the big workaround patch 177518, one of the cons is stated:
If X86 Native lib mismatch in apk is by design of some application development vendor , the patch will cause the apk run arm libraries on Intel platform even it having x86 libs, which may result in function or performance problem.
2. This AndEBench apk has this x86 native library mis-match issue, which means Intel device will run bench mark with arm native libraries instead of x86 native libraries in the apk . This definitely hurts performance because it goes through Houdini path .
Although this result is expected and it is the dilemma of this workaround, we still have consideration for this kind of problem in design of the workaround and I propose the patch to fix the performance issue for running Andebench app on Intel device.

Change-Id: Icadcc9372d7af5f84f28a33fa92bb81389ba89bd
Category: feature differentiation
Domain: AOSP.Houdini-Integration
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: zwang92 <zhenghua.wang@intel.com>